+++ fltk-sys/cfltk/cfl.h
@@ -136,6 +136,12 @@ int Fl_api_version(void);
 
 int Fl_abi_version(void);
 
+// void Fl_set_error(void (*error)(const char *, ...));
+
+// void Fl_set_warning(void (*error)(const char *, ...));
+
+// void Fl_set_fatal(void (*error)(const char *, ...));
+
 #ifdef __cplusplus
 }
 #endif