+++ fltk-sys/src/fl.rs
@@ -215,3 +215,18 @@ extern "C" {
 extern "C" {
     pub fn Fl_abi_version() -> libc::c_int;
 }
+extern "C" {
+    pub fn Fl_set_error(
+        error: ::core::option::Option<unsafe extern "C" fn(arg1: *const libc::c_char, ...)>,
+    );
+}
+extern "C" {
+    pub fn Fl_set_warning(
+        error: ::core::option::Option<unsafe extern "C" fn(arg1: *const libc::c_char, ...)>,
+    );
+}
+extern "C" {
+    pub fn Fl_set_fatal(
+        error: ::core::option::Option<unsafe extern "C" fn(arg1: *const libc::c_char, ...)>,
+    );
+}