+++ src/app.rs
@@ -10,6 +10,9 @@ use std::{
     os::raw,
 };
 
+pub type WidgetPtr = *mut fltk_sys::widget::Fl_Widget;
+
+/// The fonts associated with the application
 pub(crate) static mut FONTS: Option<Vec<String>> = None;
 
 /// Runs the event loop
@@ -361,6 +364,18 @@ where
         let data: *mut raw::c_void = a as *mut raw::c_void;
         let callback: fltk_sys::widget::Fl_Callback = Some(shim);
         fltk_sys::widget::Fl_Widget_callback_with_captures(widget.as_widget_ptr(), callback, data);
+    }
+}
+
+/// Set a widget callback using a C style API, when boxing is not desired
+pub fn set_raw_callback<W>(widget: &mut W, data: *mut raw::c_void, cb: Option<fn(WidgetPtr, *mut raw::c_void)>)
+where
+    W: WidgetExt,
+{
+    assert!(!widget.was_deleted());
+    unsafe {
+        let cb: Option<unsafe extern "C" fn(WidgetPtr, *mut raw::c_void)> = mem::transmute(cb);
+        fltk_sys::widget::Fl_Widget_callback_with_captures(widget.as_widget_ptr(), cb, data);
     }
 }