+++ src/text.rs
@@ -209,7 +209,20 @@ impl TextBuffer {
         let path = path.to_str().unwrap();
         let path = CString::new(path)?;
         unsafe {
-            match Fl_Text_Buffer_loadfile(self._inner, path.as_ptr(), 0) {
+            match Fl_Text_Buffer_loadfile(self._inner, path.as_ptr()) {
+                0 => Ok(()),
+                _ => Err(FltkError::Internal(FltkErrorKind::ResourceNotFound)),
+            }
+        }
+    }
+
+    /// Saves a buffer into a file
+    pub fn save_file(&mut self, path: &std::path::Path) -> Result<(), FltkError> {
+        assert!(!self._inner.is_null());
+        let path = path.to_str().unwrap();
+        let path = CString::new(path)?;
+        unsafe {
+            match Fl_Text_Buffer_savefile(self._inner, path.as_ptr()) {
                 0 => Ok(()),
                 _ => Err(FltkError::Internal(FltkErrorKind::ResourceNotFound)),
             }
@@ -430,6 +443,8 @@ impl TextBuffer {
     }
 
     /// Adds a modify callback
+    /// callback args:
+    /// pos: i32, inserted items: i32, deleted items: i32, restyled items: i32, deleted_text
     pub fn add_modify_callback(&mut self, cb: Box<dyn FnMut(u32, u32, u32, u32, &str)>) {
         assert!(!self._inner.is_null());
         unsafe {
@@ -467,6 +482,8 @@ impl TextBuffer {
     }
 
     /// Removes a modify callback
+    /// callback args:
+    /// pos: i32, inserted items: i32, deleted items: i32, restyled items: i32, deleted_text
     pub fn remove_modify_callback(&mut self, cb: Box<dyn FnMut(u32, u32, u32, u32, &str)>) {
         assert!(!self._inner.is_null());
         unsafe {