+++ CHANGELOG.md
@@ -5,6 +5,14 @@
 ### Changes
 - (None)
 
+## [0.8.1] - 2020-08-13
+### Changes
+- Add app::set_raw_callback() which avoids boxing callbacks.
+- Update FLTK to after commit cd26829 which adds MacOS 11.0 support.
+- Add missing docs for callback arguments.
+- Add TextBuffer::save_file().
+- Use expect() in the build script to give meaningful messages on missing dependencies.
+
 ## [0.8.0] - 2020-08-02
 ### Changes
 - Make fltk-sys no_std.